agentvision-x/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.2
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.30.0
google-generativeai>=0.8.0
